+++ bukkit/src/main/java/com/github/games647/fastlogin/bukkit/PremiumCommand.java
@@ -2,6 +2,7 @@ package com.github.games647.fastlogin.bukkit;
 
 import com.google.common.io.ByteArrayDataOutput;
 import com.google.common.io.ByteStreams;
+
 import org.bukkit.ChatColor;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;